Purge downloading toolchain logic
We do not support downloading a standard toolchain from the Internet,
because there is no good source of truth as of now.
If the user of rules_devicetree wishes to do so, they should define
their own repository, declare it with devicetree_toolchain(), then
register it themselves.

Note that with //devicetree/tests being empty, bazel test //... no
longer works. Add an empty test as a placeholder so that our CI doesn't
break.

addresses #1

Signed-off-by: Hong, Yifan <jacky8hyf@gmail.com>