Move gazelle out of the root package.
This is so that a module depending on rules_devicetree does not have to
specify an explicit dependency to bazel_gazelle. This is because
REPO.bazel specifies //:package_metadata, which requires the root package
to be always loaded. By moving it to a private package and using an alias,
bazel_gazelle won't be always needed.

addresses #1

Link: https://github.com/bazel-contrib/rules-template/issues/152
Signed-off-by: Hong, Yifan <jacky8hyf@gmail.com>